+++ app/models/category.rb
@@ -1,0 +1,13 @@
+# Allows uploaded documents to be grouped together and organized in dashboard
+# menu
+class Category < ActiveRecord::Base
+  # Concerns
+  include Deletable
+
+  # Model Validation
+  validates :name, :slug, presence: true
+  validates :slug, uniqueness: { scope: :deleted }
+  validates :slug, format: { with: /\A[a-z][a-z0-9\-]*\Z/ }
+
+  # Model Methods
+end