+++ app/models/member.rb
@@ -1,3 +1,4 @@
+# Members are directory listings that can be seen from dashboard
 class Member < ActiveRecord::Base
   # Concerns
   include Deletable