+++ db/migrate/20151020194626_create_members.rb
@@ -10,5 +10,7 @@ class CreateMembers < ActiveRecord::Migration
 
       t.timestamps null: false
     end
+
+    add_index :members, :deleted
   end
 end