+++ db/migrate/20151020213705_create_categories.rb
@@ -1,0 +1,18 @@
+class CreateCategories < ActiveRecord::Migration
+  def change
+    create_table :categories do |t|
+      t.string :name
+      t.text :description
+      t.string :top_level
+      t.string :slug
+      t.integer :position, null: false, default: 0
+      t.boolean :archived, null: false, default: false
+      t.boolean :deleted, null: false, default: false
+
+      t.timestamps null: false
+    end
+
+    add_index :categories, :archived
+    add_index :categories, :deleted
+  end
+end